seis

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -7,7 +7,7 @@
 
 void puts2(char *str)
 {
-	long int a;
+	int a;
 
 	for (a = 0 ; str[a] != '\0' ; a = a + 2)
 	{
